+++ src/Entity/Task.php
@@ -45,6 +45,11 @@ class Task
      */
     private $user;
 
+    /**
+     * @ORM\Column(type="datetime", nullable=true)
+     */
+    private $last_modification;
+
     public function __construct()
     {
         $this->createdAt = new \Datetime();
@@ -109,6 +114,18 @@ class Task
     public function setUser(?User $user): self
     {
         $this->user = $user;
+
+        return $this;
+    }
+
+    public function getLastModification(): ?\DateTimeInterface
+    {
+        return $this->last_modification;
+    }
+
+    public function setLastModification(?\DateTimeInterface $last_modification): self
+    {
+        $this->last_modification = $last_modification;
 
         return $this;
     }